+++ src/main/java/com/ebicep/warlords/pve/mobs/bosses/Illumina.java
@@ -68,7 +68,7 @@ public class Illumina extends AbstractMob implements BossMob {
             String name,
             int maxHealth,
             float walkSpeed,
-            int damageResistance,
+            float damageResistance,
             float minMeleeDamage,
             float maxMeleeDamage
     ) {
