+++ src/main/java/com/ebicep/warlords/pve/mobs/bosses/MagmaticOoze.java
@@ -67,7 +67,7 @@ public class MagmaticOoze extends AbstractMob implements BossMob {
             String name,
             int maxHealth,
             float walkSpeed,
-            int damageResistance,
+            float damageResistance,
             float minMeleeDamage,
             float maxMeleeDamage,
             int splitNumber,
@@ -93,7 +93,7 @@ public class MagmaticOoze extends AbstractMob implements BossMob {
             String name,
             int maxHealth,
             float walkSpeed,
-            int damageResistance,
+            float damageResistance,
             float minMeleeDamage,
             float maxMeleeDamage
     ) {
