+++ src/main/java/com/ebicep/warlords/pve/mobs/events/baneofimpurities/EventIllusionCore.java
@@ -27,7 +27,7 @@ public class EventIllusionCore extends AbstractEventCore {
             String name,
             int maxHealth,
             float walkSpeed,
-            int damageResistance,
+            float damageResistance,
             float minMeleeDamage,
             float maxMeleeDamage
     ) {
