+++ src/main/java/com/ebicep/warlords/pve/mobs/events/boltarobonanza/EventBoltaroShadow.java
@@ -42,7 +42,7 @@ public class EventBoltaroShadow extends AbstractMob implements BossMinionMob {
             String name,
             int maxHealth,
             float walkSpeed,
-            int damageResistance,
+            float damageResistance,
             float minMeleeDamage,
             float maxMeleeDamage,
             int split
@@ -63,7 +63,7 @@ public class EventBoltaroShadow extends AbstractMob implements BossMinionMob {
             String name,
             int maxHealth,
             float walkSpeed,
-            int damageResistance,
+            float damageResistance,
             float minMeleeDamage,
             float maxMeleeDamage
     ) {
