+++ src/main/java/com/ebicep/warlords/pve/mobs/skeleton/SkeletalArcher.java
@@ -31,7 +31,7 @@ public class SkeletalArcher extends AbstractMob implements EliteMob {
             String name,
             int maxHealth,
             float walkSpeed,
-            int damageResistance,
+            float damageResistance,
             float minMeleeDamage,
             float maxMeleeDamage,
             AbstractAbility... abilities
