+++ src/main/java/com/ebicep/warlords/pve/mobs/skeleton/SkeletalEntropy.java
@@ -33,7 +33,7 @@ public class SkeletalEntropy extends AbstractMob implements AdvancedMob {
             String name,
             int maxHealth,
             float walkSpeed,
-            int damageResistance,
+            float damageResistance,
             float minMeleeDamage,
             float maxMeleeDamage
     ) {
